Renamed files to camel case
