+++ Frontend/src/cse-ui-kit/contentBlockPopup/contentBlockPopup-Styled.tsx
