+++ Frontend/src/cse-ui-kit/contentBlockPopup/contentBlockPopup.tsx
@@ -7,7 +7,7 @@ import {
   StyledContent,
   MainText,
   BoldText
-} from './contentblock_popup-Styled';
+} from './contentBlockPopup-Styled';
 import { ReactComponent as ContentUpload } from 'src/cse-ui-kit/assets/upload-content.svg';
 
 type Props = {
